Initial commit.
This is a tiny program that uses OpenCV to detect people faces in
a photo and outputs its geometries in a way that is easy to parse
from other programs.

Note that this program has been created from some sample program
which at this time is included in the official OpenCV source code
releases —you can find it in /samples/c/facedetect.cpp.